FIX incorrect behaviour when updating colmap
Handling the position of the snake in colmap incorrectly caused at first
an issue with snake being able to collide with itself, off-by-one,
during a move even though it should only be exactly after the tail.
Secondly, after fixing this issue the snakes tail was "missing" after
moving the newly created SnakePiece on the previous tail. In other
words, since two SnakePieces momentarily shares the same gridpos it
overwrites the other ones pos.